==> moneos_2020_test/src/markdown/090_vissen/20_moneos_analyse_ankerkuil.Rmd <==
---
params:
  hoofdstuk: "090_vissen"
knit: (function(inputFile, encoding) {
        rmarkdown::render(inputFile,
                          encoding=encoding,
                          output_dir = {
                            source("../pad.R");
                            maak_pad(rmarkdown::yaml_front_matter(inputFile)$params$hoofdstuk, "output")
                          }
                          )})
title: "analyse visdata ankerkuil"
output: word_document
---


```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = FALSE, error=FALSE, warning=FALSE, message=FALSE, cache=FALSE)

```


```{r libraries}

library(tidyverse)
library(lubridate)
library(writexl)

library(INBOtheme)

```


```{r pad}

# inlezen van variabelen
# pad naar data : pad_data
# pad naar tabellen : pad_tabellen
# pad naar figuren : pad_figuren

source("../pad.R")

pad_data <- maak_pad(params$hoofdstuk, "data")
pad_figuren <- maak_pad(params$hoofdstuk, "figuren")
pad_tabellen <- maak_pad(params$hoofdstuk, "tabellen")

```


##### datasets voor aantal en gewicht:

  - _G:/Mijn Drive/INBODATA/PROJECTEN/PRJ_SCHELDE/VNSC/Rapportage_INBO/2020/090_vissen/data/ankerkuil_aantallen_2012_2018.csv_
  - _G:/Mijn Drive/INBODATA/PROJECTEN/PRJ_SCHELDE/VNSC/Rapportage_INBO/2020/090_vissen/data/ankerkuil_gewicht_2012_2018.csv_

```{r inlezen-data}

ankerkuil_aantal <- 
  read_delim(paste0(pad_data, "ankerkuil_aantallen_2012_2018.csv"),
             delim = ";")

ankerkuil_gewicht <- 
  read_delim(paste0(pad_data, "ankerkuil_gewicht_2012_2018.csv"),
             delim = ";")

data_ankerkuil <- 
  ankerkuil_aantal %>% 
  gather(soort, aantal, `tiendoornige stekelbaars`:last_col()) %>% 
  left_join(ankerkuil_gewicht %>% 
              select(-uren, -volume) %>% 
              gather(soort, gewicht, `tiendoornige stekelbaars`:last_col())) %>% 
  mutate(seizoen = recode(seizoen,
                          VJ = "voorjaar",
                          NJ = "najaar",
                          Z = "zomer")) %>% 
  mutate(seizoen = factor(seizoen, levels = c("voorjaar", "zomer", "najaar")),
         locatie = factor(locatie, levels = c(c("Doel", "Antwerpen", "Steendorp", "Branst"))))


vroegste_jaar <-
  data_ankerkuil %>% 
  pull(jaar) %>% 
  min()

laatste_jaar <-
  data_ankerkuil %>% 
  pull(jaar) %>% 
  max()

```


```{r enkel-vissen}

soorten <- 
  data_ankerkuil %>% 
  distinct(soort) %>% 
  pull(soort)

soorten_invertebraten <-
  c("wolhandkrab", "grijze garnalen", "steurgarnalen", "gammarus", "noordzeekrab", "japanse steurgarnaal", "penseelkrab")

soorten_vis <-
  setdiff(soorten, soorten_invertebraten)

data_ankerkuil <- 
  data_ankerkuil %>% 
  filter(soort %in% soorten_vis)

```


```{r aantal-en-gewicht-per-kub}

data_ankerkuil <- 
  data_ankerkuil %>% 
  mutate(aantal_per_kub = aantal/volume,
         gewicht_per_kub = gewicht/volume)

```


#### relatie aantal - gewicht

```{r relatie-aantal-gewicht, fig.height=14, fig.width=10}

# data_ankerkuil %>%
#   ggplot(aes(aantal, gewicht)) +
#   geom_point() +
#   geom_smooth(span = 10, se = FALSE) +
#   facet_wrap(~soort,
#              scales = "free",
#              ncol = 5)

data_ankerkuil %>% 
  ggplot(aes(aantal + 1, gewicht + 1)) +
  geom_point() +
  geom_smooth(span = 10, se = FALSE) +
  scale_x_log10(breaks = c(0,1,10,100,1000,10000,100000,1000000,10000000) + 1, labels = function(x) x-1) +
  scale_y_log10(breaks = c(0,1,10,100,1000,10000,100000,1000000,10000000) + 1, labels = function(x) x-1) +
  labs(x = "aantal",
       y = "gewicht") +
  facet_wrap(~soort, 
             scales = "free",
             ncol = 5)

ggsave(paste0(pad_figuren, "relatie_aantal_gewicht.jpg"))

```


#### aantal soorten

```{r aantal-soorten, fig.height=4, fig.width=8}

aantal_soorten <- 
  data_ankerkuil %>% 
  filter(aantal > 0) %>% 
  distinct(jaar, seizoen, soort) %>% 
  count(jaar, seizoen, name = "soorten") %>% 
  bind_rows(data_ankerkuil %>% 
              filter(aantal > 0) %>% 
              distinct(jaar, soort) %>% 
              count(jaar, name = "soorten") %>% 
              mutate(seizoen = "totaal")) %>% 
  mutate(seizoen = factor(seizoen, levels = c("voorjaar", "zomer", "najaar", "totaal")))

aantal_soorten %>% 
  ggplot(aes(seizoen, soorten, fill = ordered(jaar))) +
  geom_col(position = position_dodge(width = 0.925), 
           alpha = 0.8) +
  geom_text(aes(y = soorten + 3, label = soorten), 
            position = position_dodge(width = 0.925),
            size = 3) +
  labs(y = "aantal soorten",
       fill = "jaar")

ggsave(paste0(pad_figuren, "aantal_soorten.jpg"))

```


#### overzicht gevangen soorten

```{r overzicht-gevangen-soorten}

tabel_gevangen_soorten <-
  data_ankerkuil %>% 
  filter(jaar == laatste_jaar,
         aantal > 0) %>% 
  mutate(seizoen_locatie = factor(paste(seizoen, locatie, sep = "_"),
                                  levels = sapply(c("voorjaar", "zomer", "najaar"),
                                                  function(x) paste(x, c("Doel", "Antwerpen", "Steendorp", "Branst"), sep = "_")) %>%
                                    c())) %>%
  select(seizoen_locatie, soort) %>% 
  mutate(aanwezig = "x") %>% 
  spread(seizoen_locatie, aanwezig, fill = "")

knitr::kable(tabel_gevangen_soorten)

write_xlsx(list(gevangen_soorten = tabel_gevangen_soorten),
           paste0(pad_tabellen, "tabellen_vissen.xlsx"))

```


#### gemiddeld aantal gevangen individuen over seizoenen

```{r gemiddeld-over-seizoenen, fig.height=4, fig.width=6}

gemiddel_over_seizoen <- 
  data_ankerkuil %>% 
  group_by(jaar, seizoen, locatie) %>%
  summarise(aantal_per_kub = sum(aantal_per_kub)) %>% 
  ungroup() %>% 
  group_by(jaar) %>%
  summarise(aantal_per_kub = mean(aantal_per_kub)) %>% 
  ungroup()

gemiddel_over_seizoen %>% 
  ggplot(aes(ordered(jaar), aantal_per_kub)) +
  geom_col(alpha = 0.8,
           width = 0.5) +
  geom_text(aes(y = aantal_per_kub + 0.3, label = round(aantal_per_kub, 2)), 
            size = 4) +
  labs(x = NULL,
       y = expression(paste("gem. aantal ind. / ", m^3)))

ggsave(paste0(pad_figuren, "gemiddeld_over_seizoenen.jpg"))

```


#### relatief aantal gevangen individuen

```{r relatief-aantal-gevangen-individuen, fig.height=6, fig.width=4}

relatief_aantal <- 
  data_ankerkuil %>% 
  filter(jaar == 2017) %>% 
  group_by(seizoen, locatie) %>%
  mutate(relatief_aantal = aantal/sum(aantal)) %>% 
  ungroup() %>% 
  group_by(soort) %>% 
  mutate(soort2 = ifelse(any(relatief_aantal >= 0.05), soort, "rest")) %>% 
  ungroup() %>% 
  group_by(seizoen, locatie, soort2) %>% 
  summarise(aantal = sum(aantal)) %>% 
  ungroup() %>% 
  rename(soort = soort2) %>% 
  mutate(soort = factor(soort, 
                        levels = unique(soort) %>% 
                          sort() %>% 
                          setdiff("rest") %>% 
                          c(., "rest")))

relatief_aantal %>% 
  ggplot(aes(x ="", y = aantal, fill = soort)) +
  geom_col(position = "fill") +
  # geom_text(aes(label = soort),
  #            size = 3) +
  coord_polar("y", start=0) +
  facet_grid(locatie ~ seizoen) +
  labs(x = NULL,
       y = NULL) +
  scale_x_discrete(breaks = NULL) +
  scale_y_continuous(breaks = NULL) +
  theme(legend.position = "bottom",
        legend.title = element_blank())


ggsave(paste0(pad_figuren, "relatief_aantal_gevangen_individuen.jpg"))

```


```{r meta-data}

meta_data <- 
  enframe(c(vroegste_jaar = vroegste_jaar,
            laatste_jaar = laatste_jaar,
            aantal_soorten = length(soorten_vis)),
          name = "naam", value = "waarde")

meta_data %>% 
  write_delim(paste0(pad_data, "vissen_meta_data.csv"),
              delim = ";")

```
