==> moneos_2020_test/src/markdown/090_vissen/30_moneos_rapport_visdata.Rmd <==
---
params:
  hoofdstuk: "090_vissen"
knit: (function(inputFile, encoding) {
        rmarkdown::render(inputFile,
                          encoding=encoding,
                          output_dir = {
                            source("../pad.R");
                            maak_pad(rmarkdown::yaml_front_matter(inputFile)$params$hoofdstuk, "output")
                          }
                          )})
title: "moneos hoofdstuk vissen"
output:
  bookdown::word_document2:
    reference_docx: ../INBO_sjabloon.docx
bibliography: references.bib
link-citations: TRUE
csl: ../ecology.csl
---



```{r setup, include=FALSE}

knitr::opts_chunk$set(echo = FALSE, error=FALSE, warning=FALSE, message=FALSE, cache=FALSE)
knitr::opts_knit$set(eval.after = "fig.cap")

```


```{r libraries}

library(tidyverse)
library(readxl)
# library(flextable)
# library(officer)
library(kableExtra)

```


```{r pad}

# inlezen van variabelen
# pad naar data : pad_data
# pad naar tabellen : pad_tabellen
# pad naar figuren : pad_figuren

source("../pad.R")

pad_data <- maak_pad(params$hoofdstuk, "data")
pad_figuren <- maak_pad(params$hoofdstuk, "figuren")
pad_tabellen <- maak_pad(params$hoofdstuk, "tabellen")

```


```{r meta_data}

meta_data <- 
  read_delim(paste0(pad_data, "vissen_meta_data.csv"),
             delim = ";")
  
for(i in 1:nrow(meta_data)){
  ##first extract the object value
  tempobj=meta_data$waarde[i]
  ##now create a new variable with the original name of the list item
  eval(parse(text=paste(meta_data$naam[i],"= tempobj")))
}

```


# Vissen

Fichenummer: FICHE S-DS-V-004a – Vissen (KRW) & S-DS-V-004b - Vissen (Vrijwilligersnetwerk)

__Jan Breine__

##	Inleiding

De meeste vissen hebben een gevarieerde levenscyclus. Gedurende hun leven doorlopen ze verschillende niveaus in het voedselweb en bevolken ze diverse ecologische niches. Estuaria vervullen verschillende functies afhankelijk van het levensstadium waarin vissen zich bevinden. Veel vissoorten gebruiken estuaria als paaihabitat [@ableReexaminationFishEstuarine2005; @vandermeulenHabitatRequirementsSpawning2014]. De kinderkamerfunctie voor jonge vis werd uitgebreid toegelicht door @elliottFishesEstuaries2008. @maesModellingMigrationOpportunities2007, @maesPoorWaterQuality2008 en @stevensOnderzoekNaarTrekvissoorten2009 gaan dieper in op de functie van estuaria als doorgangszone voor trekvissen. Estuaria zijn voedselrijk en door de diversiteit aan habitats voorzien ze voedsel voor veel jonge en adulte vissen [@baldoaMultivariateApproachFeeding2002].

Het bestuderen van de visfauna in de Zeeschelde is belangrijk om zicht te krijgen op de invulling van deze functies en om de ecologische ontwikkelingen op lange termijn in het gebied te volgen. De Zeeschelde verzamelt een belangrijk deel van de bovenstroomse vuilvrachten. Opvolging van de visstand in de Zeeschelde houdt dus ook een spiegel voor de kwaliteit van de bovenstroomse gebieden. De Europese Kaderrichtlijn Water (KRW, WFD, 2000; [@Directive2000602000]) verplicht om iedere zes jaar de ecologische toestand van oppervlaktewaterlichamen te rapporteren. Deze ecologische toestand moet beoordeeld worden met bio-indicatoren waaronder vissen. 

De visgemeenschap in de Zeeschelde wordt jaarlijks onderzocht, niet alleen omwille van de dynamiek van het systeem maar ook om het herstel te evalueren na jarenlange plaatselijke en tijdelijke zuurstofloze condities [@marisBasiswaterkwaliteit2011]. Zesjaarlijkse afvissingen, het absolute minimum voor de KRW, geven te grote lacunes om dit herstelproces te documenteren. In het begin werd ook jaarlijks op de zijrivieren gevist om er de veranderingen in de visgemeenschap op de voet te volgen. Wegens gebrek aan personeel werd in 2012 beslist om dit naar driejaarlijks te verminderen. Om de seizoenale patronen te documenteren vissen we in drie verschillende seizoenen: lente, zomer en herfst.

Sinds 2011 is er naast de reguliere fuikvisserij (gestart in 2002 [@maesOpvolgingVanHet2003]) ook ankerkuilvisserij in de Zeeschelde [@goudswaardKuilenSchietenHet2011]. Dat gebeurde in eerste instantie in Doel en Antwerpen. In 2012 voegden we er nog twee stroomopwaarts gelegen locaties, Steendorp en Branst, aan toe [@breineOpvolgingVanHet2013a] ter vervollediging van de saliniteitsgradient. Ankerkuilvisserij is goed toepasbaar in de waterkolom, fuikvisserij eerder nabij de bodem. Beide methodes zijn complementair en geven samen een vollediger beeld van de visgemeenschappen in de Zeeschelde voor de verschillende saliniteitszones. Bovenop de resultaten van het reguliere INBO meetnet krijgen we extra informatie uit de resultaten van een vrijwilligersmeetnet.

INBO fuikvangstgegevens zijn beschikbaar via de VIS databank (https://vis.inbo.be) [@brosensVISDatabaseDistribution2015]. Vrijwilligersdata en ankerkuilgegevens worden momenteel nog afzonderlijk opgeslagen tot er een aanpassing kan gebeuren aan de databank. Alle data werden aangeleverd als xlsx-bestanden aan de Scheldemonitor.


## Materiaal en Methode

de locaties bemonsterd met ankerkuil zijn weergegeven in Figuur \@ref(fig:090-figuur-locaties-ankerkuil).


```{r 090-figuur-locaties-ankerkuil, fig.cap=cap_locaties_ankerkuil}

cap_locaties_ankerkuil <- 
  paste0("De met ankerkuil bemonsterde locaties in de Zeeschelde in de period ", 
         vroegste_jaar, 
         " - ", 
         laatste_jaar, 
         ".")

knitr::include_graphics(paste0(pad_figuren, "locaties_ankerkuil.jpg"))

```

<br/>  

## Resultaten

Het totaal aantal gevangen soorten is weergegeven in Figuur \@ref(fig:090-figuur-aantal-soorten).

```{r 090-figuur-aantal-soorten, fig.cap=cap_aantal_soorten}

cap_aantal_soorten <-
  paste0("Totaal aantal gevangen soorten per seizoen en per vangstcamapgne (", 
         vroegste_jaar, 
         " - ", 
         laatste_jaar, 
         "), en het totaal aantal gevangen soorten per vangstcampagne in de Zeeschelde.")

knitr::include_graphics(paste0(pad_figuren, "aantal_soorten.jpg"))

```

<br/>  


Een overzicht van de soorten gevangen in `r laatste_jaar` is weergegeven in Tabel \@ref(tab:090-tabel-overzicht-soorten).

```{r 090-tabel-overzicht-soorten}

tabel_gevangen_soorten <-
  read_excel(paste0(pad_tabellen, "tabellen_vissen.xlsx"),
             sheet = "gevangen_soorten")

cap_gevangen_soorten <-
  paste0("Overzicht van de vissoorten gevangen tijdens de ankerkuilcampagnes in de Zeeschelde in ",
         laatste_jaar,
         ".")

knitr::kable(NULL, caption = cap_gevangen_soorten)


tabel_gevangen_soorten  %>% 
  mutate_all(~replace_na(., "")) %>% 
  mutate_all(~cell_spec(., color = "black")) %>%
  knitr::kable(booktabs = TRUE,
               align = c("l", rep("c", 12)),
               col.names = c("", rep(c("Doel", "Antw.", "Steend.", "Branst"), 3)),
               escape = FALSE) %>%
  add_header_above(c(" " = 1, "voorjaar" = 4, "zomer" = 4, "najaar" = 4)) %>% 
  kable_styling("bordered",
                latex_options = c("scale_down"),
                font_size = 15) %>%
  # kable_styling(latex_options = c("scale_down")) %>%
  save_kable("090-tabel-overzicht-soorten.png")

knitr::include_graphics("090-tabel-overzicht-soorten.png")



# tabel_namen <-
#   tabel_gevangen_soorten %>%
#   names()
# 
# typologie_gevangen_soorten <-
#   data_frame(col_keys = tabel_namen,
#              colB = tabel_namen %>%
#                str_remove("_Doel|_Antwerpen|_Steendorp|_Branst"),
#              colA = tabel_namen %>%
#                str_remove("voorjaar_|zomer_|najaar_") %>%
#                str_replace(c("Antwerpen", "Steendorp"), c("Antw.", "Steend.")))
# 
# 
# tabel_gevangen_soorten %>%
#   # flextable(col_keys = c(tabel_namen[1], "sep_1", tabel_namen[2:5], "sep_2", tabel_namen[6:9], "sep_3", tabel_namen[10:13])) %>%
#   flextable() %>%
#   set_header_df(mapping = typologie_gevangen_soorten, key = "col_keys" ) %>%
#   merge_h(part = "header") %>%
#   merge_v(part = "header", j = 1:2) %>%
#   theme_booktabs() %>%
#   font(part = "all", fontname = "Calibri") %>%
#   fontsize(part = "all", size = 8) %>%
#   align(align = "center", part = "all") %>%
#   # vline(j = c("soort", "voorjaar_Branst", "zomer_Branst"), border = fp_border(), part = "all") %>%
#   vline(j = str_subset(tabel_namen, "soort|Branst"), border = fp_border(), part = "all") %>%
#   vline_left(border = fp_border(), part = "all") %>%
#   fix_border_issues() %>%
#   empty_blanks() %>%
#   autofit(add_w = .0) %>%
#   set_caption(cap_overzicht_soorten)

```

<br/>  

Het gemiddeld aantal individuen per m^3^ water is weergegeven in Figuur \@ref(fig:090-figuur-gemiddeld-seizoenen) en het relatief aantal gevangen individuen in Figuur \@ref(fig:090-figuur-relatief-gevangen).


```{r 090-figuur-gemiddeld-seizoenen, fig.cap=cap_gemiddeld_seizoenen}

cap_gemiddeld_seizoenen <- 
  paste0("gemiddeld aantal individuen per m³ water gevangen in de Zeeschelde in de periode ", 
         vroegste_jaar, 
         " - ", 
         laatste_jaar, 
         ".")

knitr::include_graphics(paste0(pad_figuren, "gemiddeld_over_seizoenen.jpg"))

```

<br/>  

```{r 090-figuur-relatief-gevangen, fig.cap=cap_relatief_gevangen}

cap_relatief_gevangen <- 
  paste0("Het relatief aantal gevangen individuen in de Zeeschelde tijdens de ankerkuilcampagnes in het voorjaar, de zomer en het najaar van ", 
         laatste_jaar, 
         ".")

knitr::include_graphics(paste0(pad_figuren, "relatief_aantal_gevangen_individuen.jpg"))

```

<br/>  

## Referenties
